clean up data dictionary generation script
