+++ Gemfile
@@ -2,6 +2,7 @@ source 'http://rubygems.org'
 
 gem 'rake', '~> 10.4'
 gem 'nokogiri'
+gem 'rubyXL'
 
 gem 'minitest'
-gem 'minitest-ci', :git => 'git@github.com:circleci/minitest-ci.git'
+gem 'minitest-ci', git: 'git@github.com:circleci/minitest-ci.git'
